% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max_scale.R
\name{min_max_scale}
\alias{min_max_scale}
\title{Scaling by min-max}
\usage{
min_max_scale(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A numeric vector
}
\description{
Scaling by min-max
}
\examples{
x <- 1:5
min_max_scale(x)

}
